net: lib: mqtt_helper: Add Kconfig for legacy IPv4-only if needed
This option enables the previous behaviour of IPv4 only and may be
useful if a broken IPv6 implementation is encountered, e.g. resolves
IPv6 addresses but doesn't have routing. It has  been added for
backwards compatibility. Tested on Thingy:91 in both IPv4 mode,
and in dual stack. Includes documentation updates.

Signed-off-by: Sly Gryphon <sgryphon@computer.org>